create custom hook useAlertAction
